fix(chat): avoid the circular dependency causing a Temporal Dead Zone (TDZ) error (Issue #5576)
