+++ apps/chat/src/pages/api/files/upload-archive.ts
@@ -2,6 +2,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { getServerSession } from 'next-auth/next';
 
 import { prepareFileName } from '@/src/utils/app/file';
+import { authOptions } from '@/src/utils/auth/auth-options';
 import { validateServerSession } from '@/src/utils/auth/session';
 import {
   buildApiUrl,
@@ -15,8 +16,6 @@ import { getToken } from '@/src/utils/server/server';
 import { DialAIError } from '@/src/types/error';
 
 import { errorsMessages } from '@/src/constants/errors';
-
-import { authOptions } from '@/src/pages/api/auth/[...nextauth]';
 
 import FormData from 'form-data';
 import JSZip from 'jszip';
